% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rte.R
\name{calc_rte}
\alias{calc_rte}
\title{Calculate Rank-based Triage Effectiveness (RTE)}
\usage{
calc_rte(
  df,
  subgroup = NULL,
  var1 = NULL,
  var2 = NULL,
  bootstrap = FALSE,
  bootstrap_params = list(sample_percentage = 1, n_iterations = 2000, distribution_span =
    0.95),
  n_workers = parallel::detectCores() - 1,
  verbose = FALSE,
  overall_only = FALSE,
  min_loset_warning = 5,
  include_distributions = TRUE,
  check_convergence = TRUE,
  seed = NULL,
  quiet = TRUE
)
}
\arguments{
\item{df}{Data frame containing patient data (must be initialized through init())}

\item{subgroup}{Optional list for subgroup analysis - an alternative to var1/2}

\item{var1}{Optional string for first comparison variable}

\item{var2}{Optional string for second comparison variable}

\item{bootstrap}{Logical. Whether to perform bootstrap calculations. Default is FALSE.}

\item{bootstrap_params}{List of bootstrap parameters:
\itemize{
\item sample_percentage (default: 1) How large of the sample is used? 1 = full sample.
\item n_iterations (default: 2000) How many bootstrap iterations to use
\item distribution_span (default: 0.95) What is the confidence interval used?
}}

\item{n_workers}{Number of workers for parallel processing (default: detectCores() - 1)}

\item{verbose}{Logical. If TRUE, returns patient-level RTE values for further analysis including  intermediate calculation
fields from calculate_queue_metrics. Default is FALSE.}

\item{overall_only}{Logical. If TRUE, only overall metrics are returned (default: FALSE)}

\item{min_loset_warning}{Numerical, when should the function warn for low loset prevalence? (default: 5)}

\item{include_distributions}{Logical. If TRUE, returns bootstrap distributions. Default is FALSE.}

\item{check_convergence}{Logical, Default is TRUE. Will check convergence and produce convergence plots if true + bootstrap == TRUE.}

\item{seed}{Seed for reproducible bootstrapping (default: NULL - no seed)}

\item{quiet}{Logical. Default is TRUE. If false outputs the recommended/used confidence interval choice.}
}
\value{
A list containing:
\itemize{
\item results: Tibble with RTE metrics
\item metadata: List containing calculation parameters
\item verbose_df: (If verbose) Patient-level RTE values
\item bootstrap_distributions: (If bootstrap & include_distribution) bootstrap distributions from bootstrapping.
}
}
\description{
Calculates Rank-based Triage Effectiveness (RTE) metrics, which evaluate triage performance
based on patient queue positions rather than waiting times. RTE provides a patient-level
metric that complements the waiting time-based Triage Effectiveness (WTE) approach.
}
