% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_te.R
\name{calculate_overall_metrics}
\alias{calculate_overall_metrics}
\title{Calculate Overall TE Metrics}
\usage{
calculate_overall_metrics(unit_results, overall_grouping)
}
\arguments{
\item{unit_results}{Results from unit-level calculations}

\item{overall_grouping}{Grouped data frame for overall calculations}
}
\value{
Tibble with overall TE metrics
}
\description{
Calculate Overall TE Metrics
}
\details{
Calculates overall metrics using weighted means based on
number of LOSET patients in each unit.
}
\keyword{internal}
