% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_wte.R
\name{calculate_unit_metrics}
\alias{calculate_unit_metrics}
\title{Calculate Unit-Level TE Metrics}
\usage{
calculate_unit_metrics(grouping)
}
\arguments{
\item{grouping}{Grouped data frame including unit and comparison variables}
}
\value{
Tibble with unit-level TE metrics
}
\description{
Calculate Unit-Level TE Metrics
}
\details{
Calculates unit-level metrics including:
\itemize{
\item Patient counts and LOSET prevalence
\item Mean waiting times
\item OTE, TTE, and BTTE metrics
\item Sensitivity and specificity
}
}
\keyword{internal}
