% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_te.R
\name{combine_results}
\alias{combine_results}
\title{Combine Base Results with Bootstrap Results}
\usage{
combine_results(base_results, bootstrap_results, var1, var2, subgroup)
}
\arguments{
\item{base_results}{Tibble with base TE calculations}

\item{bootstrap_results}{List containing bootstrap metrics and distributions}

\item{var1}{If applied, var1 is added.}

\item{var2}{If applied, var2 is added.}

\item{subgroup}{If applied, the subgroup chain is added.}
}
\value{
List with combined results and metadata
}
\description{
Combine Base Results with Bootstrap Results
}
\details{
Merges base calculations with bootstrap results and creates
metadata about the calculation process.
}
\keyword{internal}
