% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rte.R
\name{combine_rte_results}
\alias{combine_rte_results}
\title{Combine RTE Base Results with Bootstrap Results}
\usage{
combine_rte_results(base_results, bootstrap_metrics)
}
\arguments{
\item{base_results}{Tibble with base RTE calculations}

\item{bootstrap_metrics}{Tibble with bootstrap metrics}
}
\value{
Combined results tibble
}
\description{
Combine RTE Base Results with Bootstrap Results
}
\details{
Merges direct calculation results with bootstrap metrics.
}
\keyword{internal}
