% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_helper_classes.R
\name{create_batches}
\alias{create_batches}
\title{Create Batches from Segments}
\usage{
create_batches(segments, n_batches)
}
\arguments{
\item{segments}{A list of data frame segments created by the create_segments function.}

\item{n_batches}{Integer. The number of batches to create, typically equal to the number of available cores.}
}
\value{
A list of batches, where each batch contains one or more segments.
}
\description{
This function organizes segments of patient data into batches of approximately equal size.
Batching is used to distribute the workload evenly across available processing cores for
efficient parallel computation.
}
\details{
The function aims to create batches of roughly equal total size (in terms of number of patients)
to ensure balanced workload distribution in parallel processing. This approach optimizes
the use of computational resources during simulation.
}
\keyword{internal}
