% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_helper_classes.R
\name{create_segments}
\alias{create_segments}
\title{Create Segments from Patient Data}
\usage{
create_segments(df, n_workers = detectCores() - 1)
}
\arguments{
\item{df}{A data frame containing patient data, including ID, unit, arrival times, and resolve times.}

\item{n_workers}{Integer. Number of cores to use for parallel processing. Default is (detectCores() - 1).}
}
\value{
A list of data frames, where each data frame represents a segment of continuous queue activity.
}
\description{
This function divides the input patient data into segments based on continuous queue periods.
Each segment represents a period where the queue starts at 0 and ends at 0, allowing for
independent processing of these time periods.
}
\details{
The function identifies periods where the queue size goes from 0 to 1, marking the start of a new segment.
It then groups patient data into these segments, which can be processed independently in parallel simulations.
This segmentation is crucial for maintaining chronological integrity while enabling parallelization.
}
\keyword{internal}
