% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Package Attachment Hook}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{Character string giving the library where the package is installed}

\item{pkgname}{Character string giving the name of the package}
}
\description{
Displays startup messages when the package is attached.
Shows version information and parallel processing configuration.
}
\keyword{internal}
