% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize the Data and Settings}
\usage{
init(
  Df,
  start = NULL,
  stop = NULL,
  Logg_level = 0,
  time_critical_prio = c(1, 2)
)
}
\arguments{
\item{Df}{Data frame containing the data.}

\item{start}{Optional. The start datetime for calculating arrival and resolve minutes. If not provided, defaults to the earliest arrival time in the data frame.}

\item{stop}{Optional. The stop datetime for calculating arrival and resolve minutes. If not provided, defaults to the latest resolve time in the data frame.}

\item{Logg_level}{An integer indicating the logging level. 0 logs only errors, 1 logs successes as well.}

\item{time_critical_prio}{1 or more priorities, specified within c() that will be counted as time critical when calculated sensitivity/specificity.}
}
\value{
A data frame with validated and modified fields.
}
\description{
This function initializes the data frame and settings for the triage efficacy calculations.
It ensures that the required columns are present, validates the data, casts variables to their appropriate types,
and adds calculated fields. If 'start' and 'stop' are not provided, they default to the earliest arrival time
and the latest resolve time in the data frame, respectively. It also logs success and error messages based on the
specified logging level.
}
\details{
Of note is that we "fix" the datetimes within the function due to an error in R/lubridate that treats hms 00:00:00
as non existing, which means that they later are cast to just a date wich returns NA on ymd_hms casting.
}
\examples{
\dontrun{
init(Df = your_data_frame, start = start_date, stop = stop_date)
init(Df = your_data_frame)  # Automatically uses the earliest and latest times in the data frame
}
}
