% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{init_progressr}
\alias{init_progressr}
\title{Initialize Progress Reporting Handlers}
\usage{
init_progressr()
}
\description{
Sets up progress reporting handlers if progress reporting is enabled in package options.
}
\keyword{internal}
