% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_helper_classes.R
\name{run_simulation}
\alias{run_simulation}
\title{Run Simulation on a Single Segment}
\usage{
run_simulation(segment, tte = TRUE, btte = FALSE)
}
\arguments{
\item{segment}{A data.table representing a segment of patient data.}

\item{tte}{Logical. If TRUE, calculates Theoretical Triage Efficacy (TTE).}

\item{btte}{Logical. If TRUE, calculates Binary Theoretical Triage Efficacy (BTTE).}
}
\value{
A data.table with calculated wait times and resolve times.
}
\description{
This function performs the core simulation logic for calculating Theoretical Triage Efficacy (TTE)
and Binary Theoretical Triage Efficacy (BTTE) for a single segment of patient data.
}
\details{
This function simulates patient flow through a healthcare system in a deterministic manner,
strictly adhering to priority order and wait time within priority. It's optimized for performance
using data.table operations. It is used both by sim_te and sim_heat.
}
\keyword{internal}
