% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{trieff-package}
\alias{trieff}
\alias{trieff-package}
\title{Internal package utilities and initialization}
\description{
A comprehensive toolkit for analyzing and evaluating triage effectiveness in emergency departments and other contexts. The package provides methods for calculating theoretical and observed triage effectiveness, simulating triage, and visualizing the results through dumbell plots and heatmaps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://andrejohanssonlund.github.io/TriEff/index.html}
  \item Report bugs at \url{https://github.com/AndreJohanssonLund/trieff/issues}
}

}
\author{
\strong{Maintainer}: André Johansson \email{andre.johansson@med.lu.se} (\href{https://orcid.org/0000-0002-3500-4832}{ORCID})

}
\keyword{internal}
