% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_wte.R
\name{validate_results}
\alias{validate_results}
\title{Validate Results and Generate Warnings}
\usage{
validate_results(results, min_loset_warning)
}
\arguments{
\item{results}{Combined unit and overall results}

\item{min_loset_warning}{Least number of LOSET cases that should exists or the function will warn.}
}
\value{
Invisible NULL, generates warnings if issues found
}
\description{
Validate Results and Generate Warnings
}
\details{
Performs validation checks on calculated results:
\itemize{
\item Low LOSET counts
\item Extreme TE values
}
}
\keyword{internal}
