% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rte.R
\name{validate_rte_results}
\alias{validate_rte_results}
\title{Validate RTE Results and Generate Warnings}
\usage{
validate_rte_results(results, min_loset_warning, quiet)
}
\arguments{
\item{results}{Combined unit and overall results}

\item{min_loset_warning}{Least number of LOSET cases that should exist or the function will warn.}
}
\value{
Logical indicating if non-normality was detected in any unit
}
\description{
Validate RTE Results and Generate Warnings
}
\details{
Performs validation checks on calculated results:
\itemize{
\item Low LOSET counts
\item Units with too few valid RTE measurements
\item Potential non-normal distributions
}
}
\keyword{internal}
